backend/pyproject.toml
[tool.poetry]
name = "ag-advisor-map"
version = "0.2.0"
description = "Agricultural advice application"
authors = [
    "Michael <heronalps@gmail.com>",
    "Brenan <lundquist.Brenan@gmail.com>",
    "Nathaniel <thenathanielwatkins@gmail.com>",
    "Eric <ericssimpson2021@gmail.com>",
    "Dawson <dawson.whipple@gmail.com>",
    "Logan <loganbugenig01@gmail.com>"
]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "*"
uvicorn = "*"
httpx = "*"
python-dotenv = "*"
pydantic-settings = "*"
langchain-community = "*"
langchain-openai = "*"
langchain = "*"
langgraph-cli = "*"
langgraph = "*"
openai = "*"

[tool.poetry.group.dev.dependencies]
ruff = "*"
pytest = "*"
pytest-asyncio = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C90", "N", "PL"]

[tool.ruff.lint.isort]
known-first-party = ["scripts"]